Refactor actor spawning methods and streamline actor system traits (#48)
* Refactor actor spawning methods and streamline actor system traits

- Removed deprecated methods from `ActorSystemCoreExt` and `ActorSystemAdvancedExt` to simplify the API.
- Consolidated actor spawning logic into a single internal method, enhancing clarity and reducing redundancy.
- Updated usage of actor spawning in tests and examples to utilize the new builder pattern for improved readability and maintainability.
- Adjusted imports and module structure for better organization and consistency across the codebase.

* Refactor actor reference handling and enhance message serialization options

- Simplified the handling of lazy actor references by delegating message sending to resolved references, reducing complexity and avoiding recursion.
- Introduced a new `Format` enum for message serialization, supporting binary and JSON formats, with auto-detection capabilities for improved flexibility.
- Updated the `SpawnOptions` and `ResolveOptions` structures to utilize default constructors, enhancing code clarity and consistency.
- Improved error handling in the `SpawnBuilder` for actor name validation, ensuring clearer feedback on invalid paths.
- Adjusted documentation and examples to reflect the new default options and serialization methods.

* Refactor ActorLifecycle to use ActorId for watcher management

- Updated the ActorLifecycle struct to replace string-based identifiers with ActorId for both watchers and targets, enhancing type safety and consistency.
- Modified watch and unwatch methods to accept ActorId parameters, streamlining the registration and removal of watchers.
- Adjusted related methods and tests to reflect the new ActorId usage, ensuring proper functionality and improved clarity in actor relationships.
- Enhanced logging to utilize ActorId for better traceability during watch operations.

* Enhance Actor ID handling and update dependencies

- Introduced UUID-based ActorId and NodeId, replacing previous numeric identifiers for improved uniqueness and traceability.
- Updated ActorAddress to support new UUID format, ensuring backward compatibility with legacy formats.
- Refactored tests and system components to utilize the new ActorId structure, enhancing clarity and consistency across the codebase.
- Added `uuid` dependency to the project for UUID generation and handling.
- Modified Justfile to include necessary build dependencies for Python and C++ integration.

* Enhance actor proxies and refactor message handling

- Introduced SystemActorProxy and PythonActorServiceProxy for direct method calls, improving interaction with system and service actors.
- Refactored BucketStorage and TopicBroker to utilize remote method support, streamlining actor communication.
- Updated tests to validate new proxy methods and ensure proper functionality across various scenarios.
- Enhanced message serialization and error handling for improved robustness and clarity in actor interactions.

* Refactor StorageManager to utilize remote methods and improve message handling

- Converted StorageManager from an Actor to a remote class, enabling direct method calls.
- Replaced message-based communication with explicit method calls for bucket and topic management.
- Updated related methods to enhance clarity and maintainability, including get_bucket, get_topic, list_buckets, and get_stats.
- Adjusted tests to validate the new proxy method interactions, ensuring proper functionality across scenarios.

+++ crates/pulsing-actor/src/metrics/mod.rs
@@ -321,7 +321,7 @@ impl Default for MetricsRegistry {
 /// System-level metrics collected from SystemActor
 #[derive(Debug, Clone, Default)]
 pub struct SystemMetrics {
-    pub node_id: u64,
+    pub node_id: u128,
     pub actors_count: usize,
     pub messages_total: u64,
     pub actors_created: u64,
